Fix typing issues for ScrollPosition
